Move a few classes in LocalHook into folder
